+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v1.5.0]
+
+### Added
+- 0034/0046: Dynamic Media Download modal and image presets datasource
+- 0147: Asset Details Video component added to provide in-page video playback on Asset Details pages.
+
+### Fixed
+
+- 0126: Updated the include of AEM Responsive Grid's grid_base.less to a singular file copied into the Asset Share Commons codebase to support AEM 6.3.1 and AEM 6.4 in the same package.   
+- 0156: Asset cart does not populate correctly in AEM 6.4.0.
+- 0149: The Metadata Properties datasource now includes multi-value text widgets defined on the AEM Assets Metadata Schemas.
+
 ## [v1.4.0]
 
 ### Changed